aider: Update image paths in LandingPage.js to correct the paths for images located in the public directory.
# Aider chat conversation:

USER: the image path still seems to be incorrect because the images are not displaying this is the complete path for one of the images  C:\Users\Adoonl\Desktop\soccerTriviaQuiz\client\public\guessThe.svg

ASSISTANT: Given the complete path to the image, it seems that the images are located in the `public` directory of your React application. The correct path to use in your React components should be relative to the `public` directory and should not include `/public` in the path since React's build process will correctly reference assets from the `public` directory using a root-relative path (i.e., `/`).

Here are the diffs for the changes to the `LandingPage.js` file to correct the image paths:

```diff
--- client/src/components/LandingPage.js
+++ client/src/components/LandingPage.js
@@ ... @@
 const categories = [
-  { name: 'Managers', image: '/public/managers.svg',  class: 'managers'},
-  { name: 'Challenge', image: '/public/challenge.svg',  class: 'challenge' },
-  { name: 'Clubs', image: '/public/clubs.svg',  class: 'clubs' },
-  { name: 'International Teams', image: '/public/internationalTeams.svg', class: 'international-teams' },
-  { name: 'Soccer Geography', image: '/public/soccer-geography.svg',  class: 'soccer-geography' },
-  { name: 'Guess The Player', image: '/public/guess-the-player.svg',  class: 'guess-the' },
-  { name: 'Players', image: '/public/players.svg',  class: 'players' },
+  { name: 'Managers', image: '/managers.svg',  class: 'managers'},
+  { name: 'Challenge', image: '/challenge.svg',  class: 'challenge' },
+  { name: 'Clubs', image: '/clubs.svg',  class: 'clubs' },
+  { name: 'International Teams', image: '/internationalTeams.svg', class: 'international-teams' },
+  { name: 'Soccer Geography', image: '/soccer-geography.svg',  class: 'soccer-geography' },
+  { name: 'Guess The Player', image: '/guessThe.svg',  class: 'guess-the' },
+  { name: 'Players', image: '/players.svg',  class: 'players' },
   // ... other categories
 ];
@@ ... @@
           {categories.map((category, index) => (
             <Link to="/main" className={`category-card ${category.class}`} key={index}>
-              <img src={`/public/${category.image}`} alt={category.name} className="category-image" />
+              <img src={`${category.image}`} alt={category.name} className="category-image" />
               <h3>{category.name}</h3>
             </Link>
           ))}
```

These changes will update the `src` attribute of the `img` tags to use the correct paths for the images located in the `public` directory.